+++ HAARP/Samplers/SineSampler.cs
@@ -10,39 +10,39 @@ namespace HAARP.Samplers
         /// <summary>
         /// The frequency in Hertz.
         /// </summary>
-        public float Frequency { get; set; } = 1000;
+        public double Frequency { get; set; } = 1000.0;
 
         /// <summary>
         /// The amplitude of the wave.
         /// </summary>
-        public float Amplitude { get; set; } = 0.5f;
+        public double Amplitude { get; set; } = 0.5;
 
         /// <summary>
         /// The phase offset of the wave.
         /// </summary>
-        public float Phase { get; } = 0.0f;
+        public double Phase { get; } = 0.0f;
 
         /// <summary>
         /// The DC (vertical) offset of the wave.
         /// </summary>
-        public float DCOffset { get; set; } = 0.0f;
+        public double DCOffset { get; set; } = 0.0f;
 
         /// <summary>
         /// The spectral tilt of the wave.
         /// </summary>
-        public float SpectralTilt { get; set; } = 0.0f;
+        public double SpectralTilt { get; set; } = 0.0f;
 
         /// <summary>
         /// The nyquist frequency to use for spectral tilting.
         /// </summary>
-        public float SpectralUpperBound { get; set; } = 22050.0f;
+        public double SpectralUpperBound { get; set; } = 22050.0;
 
         public SineSampler(Synthesizer synth) : base(synth)
         {
             
         }
 
-        public SineSampler(Synthesizer synth, float frequency, float amplitude, float phase = 0.0f, float tilt = 0.0f, float dcOffset = 0.0f) : base(synth)
+        public SineSampler(Synthesizer synth, double frequency, double amplitude, double phase = 0.0f, double tilt = 0.0f, double dcOffset = 0.0f) : base(synth)
         {
             Frequency = frequency;
             Amplitude = amplitude;
@@ -52,10 +52,10 @@ namespace HAARP.Samplers
 	        SpectralTilt = tilt;
         }
 
-        public override void Update(ref float sample)
+        public override void Update(ref double sample)
         {
             state = (state + synth.TimeStep * Frequency) % 1.0f;
-            sample += ((float)Math.Sin(state * FullPhase) * Amplitude).Tilt(Frequency, SpectralTilt, SpectralUpperBound) + DCOffset;
+            sample += (Math.Sin(state * FullPhase) * Amplitude).Tilt(Frequency, SpectralTilt, SpectralUpperBound) + DCOffset;
         }
     }
 }
